requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiogram-calendar==0.5.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
docx==0.2.4
et-xmlfile==1.1.0
frozenlist==1.4.1
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
lxml==5.1.0
magic-filter==1.0.12
multidict==6.0.4
mysql-connector-python==8.3.0
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.0
pillow==10.2.0
postgres==4.0
protobuf==4.25.2
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyparsing==3.1.1
python-aiml==0.9.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
yarl==1.9.4
